feat: Add support for reporting on warning and error levels
The plugin will now report the severity according to the severity defined in the Dependency Cruiser configuration.

feat: Add support for configuration

It's now possible to configure the location of the dependency cruiser config file and the baseline file via ESLint's settings feature.

fix: Fix issues with loading TS config

Typescript config should now be correctly loaded, if it's specified in the Dependency Cruiser configuration